requirements.txt
fastapi
uvicorn
pyjwt
requests
razorpay
python-dotenv